Determine ranges of a polytope

The functions pol.ranges() and lim.ranges() compute the theoretical ranges of the polytope along each dimension of a given polytope \(\mathcal{P}= \{ x \in \mathbb{R}^n: Ax = B, Gx \geq H \}\).

pol.ranges(A = NULL, B = NULL, G, H)

lim.ranges(lim)

Arguments

A

A matrix corresponding to A in the description of the polytope \(\mathcal{P}\).

B

A numeric vector corresponding to B in the description of the polytope \(\mathcal{P}\).

G

A matrix corresponding to G in the description of the polytope \(\mathcal{P}\).

H

A numeric vector corresponding to H in the description of the polytope \(\mathcal{P}\).

lim

A list with four components A, B, G and H representing the polytope.

Value

A \(n \times 3\) matrix, with \(n\) the dimension of the polytope. The three columns of the matrix correspond respectively to the minimum, maximum and range along each dimension of the polytope.

Examples

# Create a lim object from a Description file
DF <- system.file("extdata", "DeclarationFileBOWF-short.txt", package = "samplelim")
BOWF <- df2lim(DF)
pol.ranges(A = BOWF$A, B = BOWF$B, G = BOWF$G, H = BOWF$H)
#>               min        max      range
#>  [1,] 114.4444444 860.000000 745.555556
#>  [2,]   0.8294294  19.140679  18.311249
#>  [3,]   9.7687500  91.874002  82.105252
#>  [4,]   0.0000000  27.892545  27.892545
#>  [5,]   0.0000000 848.497039 848.497039
#>  [6,]   5.7222222 258.000000 252.277778
#>  [7,]   5.7222222 430.000000 424.277778
#>  [8,]  35.8187500 170.623147 134.804397
#>  [9,]   1.3768528 136.181250 134.804397
#> [10,]   0.0000000 128.291897 128.291897
#> [11,]   0.0000000 128.291897 128.291897
#> [12,]   0.0000000 506.793943 506.793943
#> [13,]   0.0000000 308.624517 308.624517
#> [14,]   0.0000000 529.070600 529.070600
#> [15,]   0.1658859  49.222139  49.056253
#> [16,]   0.3317718  82.036898  81.705126
#> [17,]   0.8294294  19.140679  18.311249
#> [18,]   0.0000000  38.690056  38.690056
#> [19,]   0.0000000  16.407380  16.407380
#> [20,]   0.0000000  77.284275  77.284275
#> [21,]  18.2350000 195.710163 177.475163
#> [22,]  11.7225000 110.248803  98.526303
#> [23,]   0.8294294  19.140679  18.311249
#> [24,]   0.3968214  77.088716  76.691894
#> [25,]   3.1901131  31.901131  28.711018
#> [26,]   0.6380226  31.901131  31.263108
#> [27,]   0.6380226   6.380226   5.742204
#> [28,]   0.0000000   4.017179   4.017179